share common code
