+++ src/create-code/lib/db.ts
@@ -1,42 +1,9 @@
 import {Pool, QueryConfig, QueryResult} from 'pg';
-import {logError, logInfo} from './log';
-import {ReferralCreatedEvent} from "../lambda/models";
+import {ReferralCreatedEvent} from "./models";
+import {handleQuery} from "../../lib/db";
 
-export type DBConfig = {
-    username: string,
-    url: string,
-    password: string
-};
-
-// just the most useful bits of the QueryResult type from the pg library
-type QueryResultPartial = {
-    command: string,
-    rowCount: number,
-    rows: any[]
-}
-
-export type QuerySuccess = { success: QueryResultPartial };
-export type QueryError = { error: string };
-export type QueryRes = QuerySuccess | QueryError
-
-export function createDatabaseConnectionPool(dbConfig: DBConfig): Pool {
-    const match = dbConfig.url.match(/\/\/(.*)\/(.*)/);
-    if (match !== null) {
-        const [_, host, database] = match;
-        return new Pool({
-            host,
-            database,
-            user: dbConfig.username,
-            password: dbConfig.password,
-            port: 5432,
-        });
-    }
-
-    throw new Error(`Could not parse DB config ${JSON.stringify(dbConfig)}`);
-}
-
-export function referralCreatedEventToQueryConfig(record: ReferralCreatedEvent): QueryConfig {
-    return {
+export function writeReferralCode(record: ReferralCreatedEvent, pool: Pool): Promise<QueryResult> {
+    const query: QueryConfig = {
         text: `
             INSERT INTO contribution_referral_codes(
                 braze_uuid,
@@ -55,31 +22,6 @@ export function referralCreatedEventToQueryConfig(record: ReferralCreatedEvent):
             record.campaignId
         ]
     };
-}
 
-function queryToString(queryConfig: QueryConfig): string {
-    return `Query: ${JSON.stringify(queryConfig)}`;
-}
-
-export function handleQuery(
-    queryPromise: Promise<QueryResult>,
-    queryConfig: QueryConfig,
-): Promise<QueryRes> {
-    return queryPromise
-        .then((result: QueryResult) => {
-            logInfo(`${queryToString(queryConfig)}. Affected ${result.rowCount} row(s): `, result.rows);
-            return {
-                success: {
-                    command: result.command,
-                    rowCount: result.rowCount,
-                    rows: result.rows,
-                }
-            };
-        })
-        .catch((err: Error) => {
-            logError(`${queryToString(queryConfig)}. Error executing query`, err, err.stack);
-            return ({
-                error: err.message
-            });
-        });
+    return handleQuery(pool.query(query), query);
 }
