--- src/create-code/lib/stage.ts
@@ -1,7 +1,0 @@
-export function isRunningLocally(): boolean {
-    return process.env.Stage === 'DEV';
-}
-
-export function isProd(): boolean {
-    return process.env.Stage === 'PROD';
-}
